refactor:  code clean up
