+++ package.json
@@ -77,8 +77,8 @@
         "husky": "^9.1.7",
         "lint-staged": "^16.1.0",
         "mongoose": "^8.0.0",
-        "prisma": "^5.0.0",
         "prettier": "^3.5.3",
+        "prisma": "^5.0.0",
         "type-fest": "^4.41.0",
         "typedoc": "^0.28.5",
         "typedoc-github-theme": "^0.3.0",
@@ -97,7 +97,10 @@
     },
     "dependencies": {
         "@faker-js/faker": "^9.8.0",
-        "@tool-belt/type-predicates": "^1.4.0"
+        "@tool-belt/type-predicates": "^1.4.0",
+        "@vitest/coverage-v8": "^3.2.3",
+        "prettier-plugin-jsdoc": "^1.3.2",
+        "rimraf": "^6.0.1"
     },
     "packageManager": "pnpm@10.11.0+sha512.6540583f41cc5f628eb3d9773ecee802f4f9ef9923cc45b69890fb47991d4b092964694ec3a4f738a420c918a333062c8b925d312f42e4f0c263eb603551f977"
 }
