+++ src/persistence-adapter.ts
